fix: move start tour to help center
